split test_web.py into smaller files
there's more to be done, but this is a start
