+++ src/allmydata/test/web/common.py
@@ -1,0 +1,6 @@
+
+unknown_rwcap = u"lafs://from_the_future_rw_\u263A".encode('utf-8')
+unknown_rocap = u"ro.lafs://readonly_from_the_future_ro_\u263A".encode('utf-8')
+unknown_immcap = u"imm.lafs://immutable_from_the_future_imm_\u263A".encode('utf-8')
+
+FAVICON_MARKUP = '<link href="/icon.png" rel="shortcut icon" />'
